(feature) Deprecate JS logic ; Switch to Ergo-level dispatch

+++ packages/cicero-engine/test/data/saft/test/logic.js
@@ -53,7 +53,7 @@ describe('Logic', () => {
             state.$class = 'org.accordproject.contract.State';
             const result = await engine.execute(clause, request, state, false);
             result.should.not.be.null;
-            result.response.tokenAmount.should.equal(100);
+            result.response.tokenAmount.doubleValue.should.equal(100);
             result.response.tokenAddress.should.equal("Daniel Charles Selman");
         });
     });
